+++ src/components/SolicitacaoUnificada/SolicitacaoUnificada.jsx
@@ -15,11 +15,11 @@ import { Modal } from "react-bootstrap";
 import { Collapse } from "react-collapse";
 import { Stand } from "react-burgers";
 import { required, maxValue } from "../../helpers/fieldValidators";
-import SelecionaTempoPasseio from "../SolicitacaoDeKitLanche/TourRequestCheck";
-import SelecionaKitLancheBox from "../SolicitacaoDeKitLanche/SelecionaKitLancheBox";
+import SelecionaTempoPasseio from "../Shareable/KitLanche/SelecionaTempoPasseio/SelecionaTempoPasseio";
+import SelecionaKitLancheBox from "../Shareable/KitLanche/SelecionaKitLancheBox/SelecionaKitLancheBox";
 import CardMatriculados from "../Shareable/CardMatriculados";
 import TabelaHistoricoLotes from "../Shareable/TabelaHistoricoLotes";
-import { adapterEnumKits } from "../SolicitacaoDeKitLanche/helper";
+import { extrairKitsLanchesParaCards } from "../Shareable/KitLanche/helper";
 import { kitLanches } from "../../services/solicitacaoDeKitLanche.service";
 import "../Shareable/style.scss";
 import "./style.scss";
@@ -224,7 +224,7 @@ class SolicitacaoUnificada extends Component {
     kitLanches()
       .then(response => {
         this.setState({
-          enumKits: adapterEnumKits(response)
+          enumKits: extrairKitsLanchesParaCards(response)
         });
       })
       .catch(error => {
@@ -449,7 +449,7 @@ class SolicitacaoUnificada extends Component {
       res => {
         if (res.status === HTTP_STATUS.OK) {
           toastSuccess("Solicitação Unificada enviada com sucesso!");
-          this.resetForm();;
+          this.resetForm();
         } else if (res.status === HTTP_STATUS.BAD_REQUEST) {
           toastError("Houve um erro ao enviar a solicitação unificada");
         }
